"updated ATM"

+++ src/main/java/rocks/zipcode/atm/bank/PremiumAccount.java
@@ -5,14 +5,14 @@ package rocks.zipcode.atm.bank;
  */
 public class PremiumAccount extends Account {
 
-    private static final int OVERDRAFT_LIMIT = 100;
+    private static final double OVERDRAFT_LIMIT = 100;
 
     public PremiumAccount(AccountData accountData) {
         super(accountData);
     }
 
     @Override
-    protected boolean canWithdraw(int amount) {
+    protected boolean canWithdraw(Double amount) {
         return getBalance() + OVERDRAFT_LIMIT >= amount;
     }
 }
